# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AhaSend API v2
  description: |
    The AhaSend API v2 allows you to send transactional emails, manage domains, webhooks, routes, API keys, and view statistics.

    ## Authentication
    All API requests must be authenticated using a Bearer token in the Authorization header:
    ```
    Authorization: Bearer aha-sk-64-CHARACTER-RANDOM-STRING
    ```

    ## Scopes
    API keys have specific scopes that control access to different resources and actions:

    ### Message Scopes
    - `messages:send:all` - Send messages from any domain in the account
    - `messages:send:{domain}` - Send messages from a specific domain
    - `messages:cancel:all` - Cancel messages from any domain
    - `messages:cancel:{domain}` - Cancel messages from a specific domain
    - `messages:read:all` - Read messages from any domain
    - `messages:read:{domain}` - Read messages from a specific domain

    ### Domain Scopes
    - `domains:read` - Read all domains
    - `domains:write` - Create and update domains
    - `domains:delete:all` - Delete any domain
    - `domains:delete:{domain}` - Delete a specific domain

    ### Account Scopes
    - `accounts:read` - Read account information
    - `accounts:write` - Update account settings
    - `accounts:billing` - Access billing information
    - `accounts:members:read` - Read account members
    - `accounts:members:add` - Add account members
    - `accounts:members:update` - Update account members
...

    ### Webhook Scopes
...

...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "aha-sk-64-CHARACTER-RANDOM-STRING"
      description: API key for authentication
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional idempotency key for safe request retries. Must be a unique string for each logical request.
        Requests with the same key will return the same response. Keys expire after 24 hours.
      schema:
        type: string
        maxLength: 255
      example: "user-12345-create-domain-20240101"
  headers:
    IdempotentReplayed:
      description: |
        Indicates whether this response is replayed from a previous identical request.
        - `true`: Response was replayed from cache (duplicate request)
        - `false`: Response from original processing or error state
      schema:
        type: string
        enum: [true, false]
  responses:
    IdempotencyConflict:
      description: Request in progress - a request with this idempotency key is already being processed
      headers:
        Idempotent-Replayed:
          $ref: '#/components/headers/IdempotentReplayed'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "A request with this idempotency key is already in progress"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error description
      example:
        message: "Error message"
    SuccessResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Success message
      example:
        message: "Operation completed successfully"
...
    CreatedWebhook:
      allOf:
        - $ref: '#/components/schemas/Webhook'
        - type: object
          properties:
            secret:
              type: string
              description: Webhook signing secret. Returned only when the webhook is created.
          required:
            - secret
    CreateWebhookRequest:
      type: object
      required:
        - name
        - url
        - scope
      properties:
        name:
          type: string
          maxLength: 255
          description: Webhook name
        url:
          type: string
          format: uri
          description: Webhook URL
        enabled:
          type: boolean
          description: Whether the webhook is enabled
          default: true
        on_reception:
          type: boolean
          description: Trigger on message reception
          default: false
        on_delivered:
          type: boolean
          description: Trigger on message delivery
          default: false
        on_transient_error:
          type: boolean
          description: Trigger on transient errors
          default: false
        on_failed:
          type: boolean
          description: Trigger on permanent failures
          default: false
        on_bounced:
          type: boolean
          description: Trigger on bounces
          default: false
        on_suppressed:
          type: boolean
          description: Trigger on suppressions
          default: false
        on_opened:
          type: boolean
          description: Trigger on opens
          default: false
        on_clicked:
          type: boolean
          description: Trigger on clicks
          default: false
        on_suppression_created:
          type: boolean
          description: Trigger on suppression creation
          default: false
        on_dns_error:
          type: boolean
          description: Trigger on DNS errors
          default: false
        scope:
          type: string
          enum: [global, scoped]
          description: Webhook scope
        domains:
          type: array
          items:
            type: string
          nullable: true
          description: Domains this webhook applies to
      example:
        name: "Delivery Webhook"
        url: "https://example.com/webhook"
        scope: "global"
        enabled: true
        on_delivered: true
...
    PaginatedWebhooksResponse:
      type: object
      required:
        - object
        - data
        - pagination
      properties:
        object:
          type: string
          enum: [list]
          description: Object type identifier
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
          description: Array of webhooks
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # SMTP Credentials Schemas
...
paths:
...
  /v2/accounts/{account_id}/webhooks:
    get:
      summary: Get Webhooks
      description: |
        Returns a list of webhooks for the account

        **Query Parameters:**
        - `enabled`: Filter by enabled status
        - Event filters: `on_reception`, `on_delivered`, `on_transient_error`, `on_failed`, `on_bounced`, `on_suppressed`, `on_opened`, `on_clicked`, `on_suppression_created`, `on_dns_error`
        - `limit`: Maximum number of items to return (1-100, default: 100)
        - `cursor`: Pagination cursor for the next page
      operationId: getWebhooks
      tags:
        - Webhooks
      parameters:
        - name: account_id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            format: uuid
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: on_reception
          in: query
          description: Filter by reception event trigger
          schema:
            type: boolean
        - name: on_delivered
          in: query
          description: Filter by delivery event trigger
          schema:
            type: boolean
        - name: on_transient_error
          in: query
          description: Filter by transient error event trigger
          schema:
            type: boolean
        - name: on_failed
          in: query
          description: Filter by failure event trigger
          schema:
            type: boolean
        - name: on_bounced
          in: query
          description: Filter by bounce event trigger
          schema:
            type: boolean
        - name: on_suppressed
          in: query
          description: Filter by suppression event trigger
          schema:
            type: boolean
        - name: on_opened
          in: query
          description: Filter by open event trigger
          schema:
            type: boolean
        - name: on_clicked
          in: query
          description: Filter by click event trigger
          schema:
            type: boolean
        - name: on_suppression_created
          in: query
          description: Filter by suppression creation event trigger
          schema:
            type: boolean
        - name: on_dns_error
          in: query
          description: Filter by DNS error event trigger
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of items to return (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: after
          in: query
          description: Pagination cursor for the next page. Provide the value provided in `next_cursor` from the response.
          schema:
            type: string
        - name: before
          in: query
          description: Pagination cursor for the previous page.
          schema:
            type: string
      security:
        - BearerAuth: ["webhooks:read:all", "webhooks:read:{domain}"]
      x-code-samples:
        - lang: go
          label: AhaSend Go SDK
          source: |
            package main

            import (
              "context"
              "fmt"
              "log"

              "github.com/AhaSend/ahasend-go"
              "github.com/AhaSend/ahasend-go/api"
              "github.com/google/uuid"
            )

            func main() {
              // Create API client with authentication
              client := api.NewAPIClient(
                api.WithAPIKey("aha-sk-your-64-character-key"),
              )

              accountID := uuid.New()

              // Create context for the API call
              ctx := context.Background()

              // Call the ping endpoint
              response, httpResp, err := client.WebhooksAPI.GetWebhooks(
                ctx,
                accountID,
                api.GetWebhooksParams{
                  Limit: ahasend.Int32(100),
                },
                nil,
              )
              if err != nil {
                log.Fatalf("Error getting webhooks: %v", err)
              }

              // Check response
              if httpResp.StatusCode == 200 {
                fmt.Printf("✅ Status: %d\n", httpResp.StatusCode)
                if response != nil {
                  fmt.Printf("Found %d webhooks\n", len(response.Data))
                }
              } else {
                fmt.Printf("❌ Unexpected status code: %d\n", httpResp.StatusCode)
              }
            }
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWebhooksResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Webhook
      description: Creates a new webhook for event notifications
      operationId: createWebhook
      tags:
        - Webhooks
      parameters:
        - name: account_id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      security:
        - BearerAuth: ["webhooks:write:all", "webhooks:write:{domain}"]
      x-code-samples:
        - lang: go
          label: AhaSend Go SDK
          source: |
            package main

            import (
              "context"
              "fmt"
              "log"

              "github.com/AhaSend/ahasend-go/api"
              "github.com/AhaSend/ahasend-go/models/requests"
              "github.com/google/uuid"
            )

            func main() {
              // Create API client with authentication
              client := api.NewAPIClient(
                api.WithAPIKey("aha-sk-your-64-character-key"),
              )

              accountID := uuid.New()

              // Create context for the API call
              ctx := context.Background()

              // Call the ping endpoint
              response, httpResp, err := client.WebhooksAPI.CreateWebhook(
                ctx,
                accountID,
                requests.CreateWebhookRequest{
                  Name:   "Failures",
                  URL:    "https://mystartup.com/webhook",
                  Scope:  "global",
                  Enabled: true,
                  OnBounced: true,
                  OnTransientError: true,
                  OnFailed: true,
                },
                nil,
              )
              if err != nil {
                log.Fatalf("Error creating webhook: %v", err)
              }

              // Check response
              if httpResp.StatusCode == 201 {
                fmt.Printf("✅ Status: %d\n", httpResp.StatusCode)
                if response != nil {
                  fmt.Printf("Created webhook: %#v\n", response)
                }
              } else {
                fmt.Printf("❌ Unexpected status code: %d\n", httpResp.StatusCode)
              }
            }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created successfully
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedWebhook'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '412':
          $ref: '#/components/responses/IdempotencyPreconditionFailed'
        '422':
          $ref: '#/components/responses/IdempotencyPayloadMismatch'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v2/accounts/{account_id}/webhooks/{webhook_id}:
    get:
      summary: Get Webhook
      description: Returns a specific webhook by ID
      operationId: getWebhook
      tags:
        - Webhooks
      parameters:
        - name: account_id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            format: uuid
        - name: webhook_id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: ["webhooks:read:all", "webhooks:read:{domain}"]
      x-code-samples:
        - lang: go
          label: AhaSend Go SDK
          source: |
            package main

            import (
              "context"
              "fmt"
              "log"

              "github.com/AhaSend/ahasend-go/api"
              "github.com/google/uuid"
            )

            func main() {
              // Create API client with authentication
              client := api.NewAPIClient(
                api.WithAPIKey("aha-sk-your-64-character-key"),
              )

              accountID := uuid.New()

              // Create context for the API call
              ctx := context.Background()

              // Call the ping endpoint
              response, httpResp, err := client.WebhooksAPI.GetWebhook(
                ctx,
                accountID,
                uuid.MustParse("c5a32c40-b351-439f-8230-779daed3e42c"),
                nil,
              )
              if err != nil {
                log.Fatalf("Error getting webhook: %v", err)
              }

              // Check response
              if httpResp.StatusCode == 200 {
                fmt.Printf("✅ Status: %d\n", httpResp.StatusCode)
                if response != nil {
                  fmt.Printf("Webhook: %#v\n", response)
                }
              } else {
                fmt.Printf("❌ Unexpected status code: %d\n", httpResp.StatusCode)
              }
            }
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update Webhook
      description: |
        Updates an existing webhook

        **Note:** The webhook secret is not updatable
      operationId: updateWebhook
      tags:
        - Webhooks
      parameters:
        - name: account_id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            format: uuid
        - name: webhook_id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: ["webhooks:write:all", "webhooks:write:{domain}"]
      x-code-samples:
        - lang: go
          label: AhaSend Go SDK
          source: |
            package main

            import (
              "context"
              "fmt"
              "log"

              "github.com/AhaSend/ahasend-go"
              "github.com/AhaSend/ahasend-go/api"
              "github.com/AhaSend/ahasend-go/models/requests"
              "github.com/google/uuid"
            )

            func main() {
              // Create API client with authentication
              client := api.NewAPIClient(
                api.WithAPIKey("aha-sk-your-64-character-key"),
              )

              accountID := uuid.New()

              // Create context for the API call
              ctx := context.Background()

              // Call the ping endpoint
              response, httpResp, err := client.WebhooksAPI.UpdateWebhook(
                ctx,
                accountID,
                uuid.MustParse("c5a32c40-b351-439f-8230-779daed3e42c"),
                requests.UpdateWebhookRequest{
                  Name:      ahasend.String("Failures"),
                  URL:       ahasend.String("https://mystartup.com/webhook"),
                  Enabled:   ahasend.Bool(true),
                  OnBounced: ahasend.Bool(true),
                },
              )
              if err != nil {
                log.Fatalf("Error updating webhook: %v", err)
              }

              // Check response
              if httpResp.StatusCode == 200 {
                fmt.Printf("✅ Status: %d\n", httpResp.StatusCode)
                if response != nil {
                  fmt.Printf("Updated webhook: %#v\n", response)
                }
              } else {
                fmt.Printf("❌ Unexpected status code: %d\n", httpResp.StatusCode)
              }
            }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete Webhook
      description: Deletes a webhook
      operationId: deleteWebhook
      tags:
        - Webhooks
      parameters:
        - name: account_id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            format: uuid
        - name: webhook_id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: ["webhooks:delete:all", "webhooks:delete:{domain}"]
      x-code-samples:
        - lang: go
          label: AhaSend Go SDK
          source: |
            package main

            import (
              "context"
              "fmt"
              "log"

              "github.com/AhaSend/ahasend-go/api"
              "github.com/google/uuid"
            )

            func main() {
              // Create API client with authentication
              client := api.NewAPIClient(
                api.WithAPIKey("aha-sk-your-64-character-key"),
              )

              accountID := uuid.New()

              // Create context for the API call
              ctx := context.Background()

              // Call the ping endpoint
              response, httpResp, err := client.WebhooksAPI.DeleteWebhook(
                ctx,
                accountID,
                uuid.MustParse("c5a32c40-b351-439f-8230-779daed3e42c"),
              )
              if err != nil {
                log.Fatalf("Error deleting webhook: %v", err)
              }

              // Check response
              if httpResp.StatusCode == 200 {
                fmt.Printf("✅ Status: %d\n", httpResp.StatusCode)
                if response != nil {
                  fmt.Printf("Deleted webhook: %#v\n", response)
                }
              } else {
                fmt.Printf("❌ Unexpected status code: %d\n", httpResp.StatusCode)
              }
            }
      responses:
        '200':
          description: Webhook deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # SMTP Credentials endpoints
...